Bump to /v38, libgit2 1.9.6, and Go 1.25
Migrate the module to github.com/northhalf/git2go/v38 and raise the Go
toolchain baseline to 1.25, upgrading all dependencies to their latest
releases with direct and indirect requires in separate blocks.

The libgit2 submodule moves from vendor/libgit2 to third_party/libgit2
and is pinned to v1.9.6. Go 1.14+ treats a top-level vendor/ directory
as a Go module vendor tree, which conflicts with the C submodule, so
relocating it also fixes the "inconsistent vendoring" failure under
modern Go.

The build-tag files accept libgit2 1.9.x at compile time and reject
builds with GIT_EXPERIMENTAL_SHA256 enabled, since this binding uses a
fixed 20-byte Oid. script/build-libgit2.sh is updated for libgit2 1.9
CMake options (BUILD_TESTS/BUILD_CLI, DEPRECATE_HARD=OFF to keep the
deprecated update_tips callback path source-compatible).

Restore the baseline test suite on the new toolchain and libgit2:
- wrapper.c includes <git2/sys/errors.h> for git_error_set_str.
- remote.go uses ucbool for git_fetch_options.update_fetchhead, which
  became unsigned in libgit2 1.9.
- Test fixtures: default branch head is set to master (libgit2 1.9
  rebase tests look up the local master branch), SSH test keys are
  2048-bit (Go 1.26 rejects 512-bit RSA), the credential String
  formatter avoids a recursive vet error, and the external HTTPS clone
  test is replaced by an in-process git-http-backend CGI fixture so the
  suite does not depend on public network services.

+++ credentials.go
@@ -73,7 +73,7 @@ func (t CredentialType) String() string {
 	}
 
 	if t != 0 {
-		parts = append(parts, fmt.Sprintf("CredentialType(%#x)", t))
+		parts = append(parts, fmt.Sprintf("CredentialType(%#x)", uint(t)))
 	}
 
 	return strings.Join(parts, "|")
